refactor: 티켓 booking 으로 이동
