+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/booking/repository/TicketRepository.java
@@ -1,0 +1,8 @@
+package com.pgms.coredomain.domain.booking.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.pgms.coredomain.domain.booking.Ticket;
+
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
+}
